+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/__init__.py
@@ -8,8 +8,8 @@ import azure.cli.command_modules.acs._help  # pylint: disable=unused-import
 
 
 def load_params(_):
-    import azure.cli.command_modules.acs._params  # pylint: disable=redefined-outer-name
+    import azure.cli.command_modules.acs._params  # pylint: disable=redefined-outer-name, unused-variable
 
 
 def load_commands():
-    import azure.cli.command_modules.acs.commands  # pylint: disable=redefined-outer-name
+    import azure.cli.command_modules.acs.commands  # pylint: disable=redefined-outer-name, unused-variable