+++ src/command_modules/azure-cli-interactive/azclishell/app.py
@@ -562,8 +562,24 @@ class Shell(object):
             CONFIG.load(os.path.join(azure_folder, 'az.json'))
             SESSION.load(os.path.join(azure_folder, 'az.sess'), max_age=3600)
 
-            config = Configuration()
-            self.app.initialize(config)
+            self.app.initialize(Configuration())
+
+            if '--progress' in args:
+                args.remove('--progress')
+                thread = ExecuteThread(self.app.execute, args)
+                thread.daemon = True
+                thread.start()
+                self.threads.append(thread)
+                self.curr_thread = thread
+
+                thread = ProgressViewThread(progress_view, self)
+                thread.daemon = True
+                thread.start()
+                self.threads.append(thread)
+                result = None
+
+            else:
+                result = self.app.execute(args)
 
             if '--progress' in args:
                 args.remove('--progress')