+++ src/command_modules/azure-cli-interactive/azclishell/progress.py
@@ -31,6 +31,9 @@ class ShellProgressView(ProgressViewBase):
         percent = args.get('percent', None)
         if percent:
             PROGRESS_BAR = self._format_value(message, percent)
+            if int(percent) == 1:
+                PROGRESS_BAR = None
+
         PROGRESS = message
 
     def _format_value(self, msg, percent=0.0):
@@ -64,7 +67,7 @@ def get_done():
 
 def progress_view(shell):
     """ updates the view """
-    global HEART_BEAT
+    global HEART_BEAT, DONE, PROGRESS_BAR
     _, col = get_window_dim()
     col = int(col)
     progress = get_progress_message()
@@ -78,6 +81,9 @@ def progress_view(shell):
     if PROGRESS_BAR:
         doc = u'{}:{}'.format(progress, PROGRESS_BAR)
         shell.spin_val = -1
+
+        counter = 0
+        HEART_BEAT = ''
     else:
         if progress and not DONE:
             if shell.spin_val >= 0:
@@ -87,6 +93,7 @@ def progress_view(shell):
                 len_beat = len(HEART_BEAT)
                 if len_beat > buffer_size:
                     HEART_BEAT = HEART_BEAT[len_beat - buffer_size:]
+
                 while len(HEART_BEAT) < buffer_size:
                     beat = HEART_BEAT_VALUES[_get_heart_frequency()]
                     HEART_BEAT += beat
@@ -103,6 +110,8 @@ def progress_view(shell):
         initial_document=Document(doc))
     shell.cli.request_redraw()
     if DONE:
+        DONE = False
+        PROGRESS_BAR = ''
         shell.spin_val = -1
         return True