+++ src/command_modules/azure-cli-interactive/HISTORY.rst
@@ -6,6 +6,8 @@ Release History
 0.3.2 (unreleased)
 ++++++++++++++++++
 * --progress flag
+* Removed --debug and --verbose from completions
+
 
 
 0.3.2 (2017-05-18)